# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: ELOJIM API
  description: API oficial del proyecto Fundación ELOJIM
  version: "1.0.0"
  contact:
    name: Soporte ELOJIM
    email: soporte@elojim.org
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    RegistroBase:
      type: object
      properties:
        nombreCompleto:
          type: string
          example: "Laura Gómez"
        fechaNacimiento:
          type: string
          format: date
        edad:
          type: integer
        comuna:
          type: string
        estratoSocial:
          type: string
        grupoEtnico:
          type: string
        telefonoContacto:
          type: string
        correoElectronico:
          type: string
        direccion:
          type: string
        motivacion:
          type: string
        expectativas:
          type: string
        aceptaTerminos:
          type: boolean
      required:
        - nombreCompleto
        - fechaNacimiento
        - comuna
        - estratoSocial
        - telefonoContacto
        - correoElectronico
        - aceptaTerminos
    RegistroSemilleroInnovacion:
      allOf:
        - $ref: "#/components/schemas/RegistroBase"
        - type: object
          properties:
            tipoDocumento:
              type: string
            numeroDocumento:
              type: string
            tipoVinculacion:
              type: string
            nombreEntidadVinculacion:
              type: string
            nivelEducativo:
              type: string
            participacionPrevia:
              type: boolean
            areasInteres:
              type: array
              items:
                type: string
            otrasAreas:
              type: string
            tieneProyecto:
              type: boolean
            descripcionProyecto:
              type: string
            habilidades:
              type: string
            disponibilidad:
              type: string
          example:
            nombreCompleto: "Pablo Rodríguez"
            tipoDocumento: "CE"
            numeroDocumento: "1098765432"
            comuna: "Comuna 8"
            nivelEducativo: "Tecnica_Tecnologica"
            areasInteres: ["Emprendimiento social", "Tecnología e innovación"]
            descripcionProyecto: "Una app móvil para mujeres artesanas"
            aceptaTerminos: true
...
security:
  - cookieAuth: []
paths:
...
  /api/registro/seguridad-alimentaria:
    get:
      tags: [Registro]
      summary: Listar registros de seguridad alimentaria
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro en seguridad alimentaria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombreCompleto: { type: string, example: "Jorge Castillo" }
                tipoDocumento: { type: string, example: "CC" }
                numeroDocumento: { type: string, example: "1022334455" }
                fechaNacimiento: { type: string, format: date, example: "1979-06-10" }
                edad: { type: integer, example: 45 }
                telefonoContacto: { type: string, example: "3199876543" }
                correoElectronico: { type: string, example: "jorge.castillo@example.com" }
                direccion: { type: string, example: "Carrera 60 #45-12" }
                barrio: { type: string, example: "La Floresta" }
                comuna: { type: string, example: "Comuna 11" }
                estratoSocial: { type: string, example: "E2" }
                grupoEtnico: { type: string, example: "Rom_Gitano" }
                esAgricultor: { type: boolean, example: true }
                tieneTierras: { type: boolean, example: true }
                hectareas: { type: number, format: float, example: 5.0 }
                pisoTermico: { type: string, example: "Cálido" }
                tieneCultivo: { type: boolean, example: true }
                tiposCultivo: { type: string, example: "Caña de azúcar, Sorgo" }
                participacionPrevia: { type: boolean, example: true }
                proyectosAnteriores: { type: string, example: "Red Rural 2019" }
                tieneRiego: { type: boolean, example: true }
                tieneHerramientas: { type: boolean, example: true }
                tiposHerramientas: { type: string, example: "Tractor, Arado, Machete" }
                tieneAsistenciaTecnica: { type: boolean, example: true }
                motivacion: { type: string, example: "Expandir mi finca y mejorar la comercialización." }
                tiempoSemanal: { type: string, example: "35 horas" }
                expectativas: { type: string, example: "Financiamiento y acceso a tecnología." }
                aceptaTerminos: { type: boolean, example: true }
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "401":
          description: No autenticado
        "500":
          description: Error interno

    delete:
      tags: [Registro]
      summary: Eliminar un registro de seguridad alimentaria
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno
  /api/registro/semillero-innovacion:
    get:
      tags: [Registro]
      summary: Listar registros de semillero de innovación
      responses:
        "200":
          description: Lista de registros
    post:
      tags: [Registro]
      summary: Crear registro en semillero innovacion 
      description: Crea un nuevo registro de participante para el programa Semillero Innovación.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroSemilleroInnovacion"
      responses:
        "201":
          description: Registro creado exitosamente
        "400":
          description: Error en validación o datos incompletos
        "500":
          description: Error interno
    delete:
      tags: [Registro]
      summary: Eliminar un registro de semillero de innovación
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 16
      responses:
        "200":
          description: Registro eliminado exitosamente
        "404":
          description: Registro no encontrado
        "500":
          description: Error interno
...